3-args

+++ 0x0A-argc_argv/3-mul.c
@@ -9,7 +9,7 @@
 int main(int argc, char *argv[])
 {
 	int i;
-	long int mul = 1;
+	int mul = 1;
 
 	if (argc == 1)
 	{
@@ -20,7 +20,7 @@ int main(int argc, char *argv[])
 	{
 		for (i = 1; i < argc; i++)
 			mul *= atoi(argv[i]);
-		printf("%ld\n", mul);
+		printf("%d\n", mul);
 	}
 	return (0);
 }
